+++ app/utils/server/navigraph/navdata/misc.ts
@@ -161,7 +161,7 @@ export const processNavdataAirways: NavdataProcessFunction = async ({ fullData, 
         let objectAirway = fullData.airways[key];
         let shortAirway = shortData.airways[key];
 
-        if (previousAirway && previousAirway.seqno + 10 === airway.seqno && airway.route_identifier === previousAirway.route_identifier) continue;
+        if (previousAirway && (previousAirway.seqno + 5 === airway.seqno || previousAirway.seqno + 10 === airway.seqno) && airway.route_identifier === previousAirway.route_identifier) continue;
 
         const nextAirways: typeof airways = [airway];
 
@@ -169,7 +169,7 @@ export const processNavdataAirways: NavdataProcessFunction = async ({ fullData, 
         let previousSeqno = airway.seqno;
         let nextItem = airways[k];
 
-        while (nextItem && airway.route_identifier === nextItem.route_identifier && previousSeqno + 10 === nextItem.seqno) {
+        while (nextItem && airway.route_identifier === nextItem.route_identifier && (previousSeqno + 10 === nextItem.seqno || previousSeqno + 5 === nextItem.seqno)) {
             nextAirways.push(nextItem);
             previousSeqno = nextItem.seqno;
             k++;